refactor(lsp): move remaining LSP classes into lsp package
Доперенесены оставшиеся классы протокола/жизненного цикла LSP из корня в lsp/:
AutoServerInfo (extends lsp4j ServerInfo), ParentProcessWatcher (слежение за
родительским процессом редактора), WorkDoneProgressHelper (LSP work-done
progress) и AnalyzeProjectOnStart (запуск анализа на старте сервера) — вместе
с тестами. AnalyzeProjectOnStart перенесён со своим resource-бандлом
(_ru/_en.properties): Resources грузит бандл по FQN класса, поэтому при смене
пакета ресурсы обязаны переезжать рядом.

В корне остались только bootstrap-классы: BSLLSPLauncher (Main-Class) и
BSLLSBinding (@ComponentScan-якорь).

ArchUnit: в allow-list правила про стандартные потоки обновлён FQN
ParentProcessWatcher на новый пакет.
