+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/lsp/WorkDoneProgressHelper.java
@@ -19,10 +19,8 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with BSL Language Server.
  */
-package com.github._1c_syntax.bsl.languageserver;
+package com.github._1c_syntax.bsl.languageserver.lsp;
 
-import com.github._1c_syntax.bsl.languageserver.lsp.ClientCapabilitiesHolder;
-import com.github._1c_syntax.bsl.languageserver.lsp.LanguageClientHolder;
 
 import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
