+++ src/main/resources/com/github/_1c_syntax/bsl/languageserver/lsp/AnalyzeProjectOnStart_ru.properties
